% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{makeGoodShape}
\alias{makeGoodShape}
\title{Fixes bad coordinates i z}
\usage{
makeGoodShape(z)
}
\arguments{
\item{z:}{A data frame containing x, y and id.}
}
\value{
The function returns a data frame with coordinates
}
\description{
makeGoodShape takes a set of points (x and y coordinates) and fixes them so they are ready for hexamap.
}
\examples{

library(hexamapmaker)
library(ggplot2)

## Example using a "normal" grid
x <- c(1,2,1,2,1,2,4,4)
y <- c(1,1,2,2,3,3,1,2)

id <- c("test1", "test2", "test3", "test4", "test5", "test6", "test7", "test8")

z <- data.frame(id,x,y)

ggplot(z, aes(x, y, group = id)) +
  geom_point() +  coord_fixed(ratio = 1)

zz <- hexamap(z)

ggplot(zz, aes(x, y, group = id)) +
  geom_polygon(colour="black", fill = NA) +
  coord_fixed(ratio = 1) ## Messed up

## Now use new function to fix structure:
newZ <- makeGoodShape(z)

ggplot(newZ, aes(x, y, group = id)) +
  geom_point() +  coord_fixed(ratio = 1)

zzz <- hexamap(newZ)

ggplot(zzz, aes(x, y, group = id)) +
  geom_polygon(colour="black", fill = NA) +
  coord_fixed(ratio = 1)
}

